+++ packages/bepark/game-finder/src/Factories/ClientFactory.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace BePark\GameFinder\Clients;
+namespace BePark\GameFinder\Factories;
 
+use BePark\GameFinder\Clients\AbstractClient;
 use Illuminate\Support\Collection;
 
 class ClientFactory
